# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Radar - Laravel",
        "description": "This is a Rest API built with Laravel for tracking personal flight hidstory",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register account",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "description": "Register an account",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created account"
                    },
                    "422": {
                        "description": "Validation failed. Email is not unique"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login account",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "description": "Login to an existing account",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully created account"
                    },
                    "422": {
                        "description": "Validation failed. Email is not unique"
                    }
                }
            }
        },
        "/api/refresh": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh JWT Token",
                "operationId": "86e894df0ccad042116bf937e5593629",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved a new valid JWT Token"
                    },
                    "400": {
                        "description": "Token is not provided"
                    },
                    "403": {
                        "description": "JWT Token is invalid, unable to generate new JWT Token"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}